04-microservice/bugs/requirements.txt
async-timeout==4.0.2
Brotli==1.0.9
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
Deprecated==1.2.13
escape==1.1
essential-generators==1.0
Flask==2.1.0
flask-api-cache==0.0.6
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
gevent==22.8.0
geventhttpclient==2.0.2
googleapis-common-protos==1.56.2
greenlet==1.1.3.post0
grpcio==1.49.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
memoization==0.3.1
msgpack==1.0.4
opentelemetry-api==1.13.0
opentelemetry-distro==0.34b0
opentelemetry-exporter-jaeger==1.13.0
opentelemetry-exporter-jaeger-proto-grpc==1.13.0
opentelemetry-exporter-jaeger-thrift==1.13.0
opentelemetry-instrumentation==0.34b0
opentelemetry-instrumentation-flask==0.34b0
opentelemetry-instrumentation-requests==0.34b0
opentelemetry-instrumentation-wsgi==0.34b0
opentelemetry-sdk==1.13.0
opentelemetry-semantic-conventions==0.34b0
opentelemetry-util-http==0.34b0
packaging==21.3
protobuf==3.20.3
psutil==5.9.2
pymongo==4.2.0
pyparsing==3.0.9
pyzmq==24.0.1
redis==4.3.4
requests==2.28.1
roundrobin==0.0.4
six==1.16.0
thrift==0.16.0
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.5.0
